Create Sum of Consecutive Integers to Target
In this program our goal is, given some positive integer, compute the number of possible ways to sum consecutive integers to reach this target value. For example, 5 = 2 + 3. Indeed, the inner workings of this problem reside primarily in the world of mathematics, so the variable l is assigned with purpose. To reach such a value, consider the arbitrary starting value i on the range 1 to n. Then we wish to find l such that the sum from k = 0 to k = l of i + k = n (this is a sum of l + 1 consecutive integers). Then, simplification yields a quadratic equation in l, so in order to yield a solution we the solution to the quadratic equation must be an integer. This is the idea.