Create Maximization with Projection
The following code demonstrates the power of linear algebra as an application to engineering, data science, and other scientific fields. Inspired by content presented in Sheldon Axler's "Linear Algebra Done Right," this code exploits the use of the projection onto a simpler subspace (in this case one spanned by a polynomials) in an arbitrary vector space. Axler elegantly demonstrates the effectiveness of this technique by delving into theory, of which I leave to the reader to explore.